requirements.txt
chromedriver-binary==86.0.4240.22.0
Pillow==8.0.1
selenium==3.141.0
urllib3==1.26.2
